Refactor Oracle into partials for save/clipboard/migrations
- Convert Expansion.Oracle to a partial class and add an XML doc map pointing to Systems.Save entrypoints.
- Extract clipboard import/export + debug buttons into Oracle.Clipboard.cs.
- Extract migration orchestration into Oracle.Migrations.cs.
- Extract autosave scheduling + preset sync triggers into Oracle.SaveScheduling.cs.
- Extract skill point reconciliation + tooling into Oracle.SkillPoints.cs.
- Extract preset helpers and skill tree dictionary bridge into dedicated Oracle partials.
- Update SaveSystemV11ImplementationLog with the new file map.

No intended behavior change; this is a mechanical move to reduce Oracle.cs coupling and make save failures easier to reason about.
