+++ Assets/Scripts/Expansion/Oracle.cs
@@ -31,7 +31,20 @@ using UnityEditor;
 
 namespace Expansion
 {
-    public class Oracle : SerializedMonoBehaviour
+    /// <summary>
+    /// Game-wide state/root "source of truth" MonoBehaviour.
+    /// </summary>
+    /// <remarks>
+    /// Save system components have been extracted out of this file to reduce coupling and make failures easier to
+    /// reason about.
+    /// <para>Canonical save string codec: <see cref="Systems.Save.SaveCodec"/> (prefix <c>IDB1:</c>).</para>
+    /// <para>Snapshot compaction (bits/filtering/facility normalization): <see cref="Systems.Save.SaveSnapshotBuilder"/>.</para>
+    /// <para>On-disk canonical string: <see cref="Systems.Save.SaveSystem"/> + <see cref="Systems.Save.OdinStringFileStorage"/>.</para>
+    /// <para>Legacy ES3 import/recovery: <see cref="Systems.Save.LegacyEs3Save"/>.</para>
+    /// <para>Legacy-load candidate selection: <see cref="Systems.Save.SaveLoadCandidateSelector"/>.</para>
+    /// <para>Clipboard UI entrypoints: <c>Assets/Scripts/Expansion/Oracle.Clipboard.cs</c>.</para>
+    /// </remarks>
+    public partial class Oracle : SerializedMonoBehaviour
     {
         public Button recoveryButton;
         public TMP_Text recoveryText;
@@ -75,549 +88,16 @@ namespace Expansion
         public GameObject skillsHolder;
         public List<int> listOfSkillsNotToAutoBuy = new List<int>();
 
-        //public List<int> skillAutoAssignmentList = new();
-        public bool Loaded;
-        private bool _isSaveReady;
-        private bool _autoSaveScheduled;
-        private bool _quickSaveScheduled;
-        private int _suppressPresetSyncDepth;
-	        private const float QuickSaveDelaySeconds = 0.25f;
-	        // IMPORTANT: When adding a migration step in BuildMigrationRegistry(),
-	        // you MUST also update this constant to match the new LatestVersion.
-	        // IMPORTANT: Save v7 introduces skill bitsets (see SkillIdMap/SkillBitsetUtility).
-	        private const int CurrentSaveVersion = 11;
+	        //public List<int> skillAutoAssignmentList = new();
+	        public bool Loaded;
+		        // IMPORTANT: When adding a migration step in BuildMigrationRegistry(),
+		        // you MUST also update this constant to match the new LatestVersion.
+		        // IMPORTANT: Save v7 introduces skill bitsets (see SkillIdMap/SkillBitsetUtility).
+		        private const int CurrentSaveVersion = 11;
 
-        #region SaveAndLoadFromClipboard
-
-        [TabGroup("SaveData", "Buttons")] public bool beta;
-
-        [TabGroup("SaveData", "Buttons"), Button("Unlock Dev Options (Testing)")]
-        public void UnlockDevOptionsForTesting()
-        {
-            if (saveSettings == null) saveSettings = new SaveDataSettings();
-
-            saveSettings.debugOptions = true;
-            PlayerPrefs.SetInt("debug", 1);
-            PlayerPrefs.Save();
-            if (lsm != null) lsm.SetDebug();
-            NotifyDebugOptionsChanged();
-
-            Debug.Log("Dev options unlocked for testing.");
-        }
-
-        [TabGroup("SaveData", "Buttons"), Button]
-        public void LoadFromClipboard()
-        {
-            bool previousDevOptions = saveSettings != null && saveSettings.debugOptions;
-            bool previousDoubleIp = saveSettings != null && saveSettings.doubleIp;
-            bool debugPrefUnlocked = PlayerPrefs.GetInt("debug", 0) == 1;
-            bool doubleIpPrefUnlocked = PlayerPrefs.GetInt("doubleip", 0) == 1;
-            string clipboard = GUIUtility.systemCopyBuffer;
-            if (!SaveCodec.TryDecodeSaveSettings(clipboard, out SaveDataSettings decoded))
-            {
-                Debug.LogError("LoadFromClipboard failed to decode save data. Clipboard format not recognized.");
-                return;
-            }
-
-            saveSettings = decoded;
-
-            LoadDictionaries();
-            if (!oracle.saveSettings.cheater && saveSettings.maxOfflineTime < 86400)
-                oracle.saveSettings.maxOfflineTime = 86400;
-            saveSettings.debugOptions = saveSettings.debugOptions || previousDevOptions || debugPrefUnlocked;
-            saveSettings.doubleIp = saveSettings.doubleIp || previousDoubleIp || doubleIpPrefUnlocked;
-            if (saveSettings.debugOptions) PlayerPrefs.SetInt("debug", 1);
-            if (saveSettings.doubleIp) PlayerPrefs.SetInt("doubleip", 1);
-            NotifyDebugOptionsChanged();
-            FixSkillpoints();
-            ApplyMigrations();
-            SyncAutoAssignFromSelectedPreset(runAutoAssign: false);
-            LogPresetAutoAssignState("PostMigration/Clipboard");
-            SaveInternal(true, updateQuitTime: false);
-            SceneManager.LoadScene(0);
-        }
-
-        private void LogPresetAutoAssignState(string label)
-        {
-            DysonVerseSaveData data = saveSettings?.dysonVerseSaveData;
-            if (data == null)
-            {
-                Debug.Log($"[PresetAutoAssign:{label}] save data missing");
-                return;
-            }
-
-            Debug.Log(
-                $"[PresetAutoAssign:{label}] " +
-                $"p1={FormatPresetList(data.skillAutoAssignmentIds1)} " +
-                $"p2={FormatPresetList(data.skillAutoAssignmentIds2)} " +
-                $"p3={FormatPresetList(data.skillAutoAssignmentIds3)} " +
-                $"p4={FormatPresetList(data.skillAutoAssignmentIds4)} " +
-                $"p5={FormatPresetList(data.skillAutoAssignmentIds5)}");
-        }
-
-        private void SyncAutoAssignFromSelectedPreset(bool runAutoAssign)
-        {
-            DysonVerseSaveData data = saveSettings?.dysonVerseSaveData;
-            if (data == null) return;
-            int presetIndex = Mathf.Clamp(data.selectedPreset, 1, 5);
-            List<string> presetIds = GetPresetAutoAssignmentSkillIds(presetIndex);
-            if (presetIds == null || presetIds.Count == 0) return;
-            SetAutoAssignmentSkillIds(new List<string>(presetIds));
-            if (runAutoAssign && _gameManager != null)
-            {
-                _gameManager.AutoAssignSkillsInvoke();
-            }
-        }
-
-        private static string FormatPresetList(List<string> ids)
-        {
-            if (ids == null) return "null";
-            if (ids.Count == 0) return "empty";
-            return $"{ids.Count}:[{string.Join(",", ids)}]";
-        }
-
-        private void FixSkillpoints()
-        {
-            if (saveSettings.hasFixedIP) return;
-            long totalEarned = prestigeData.permanentSkillPoint + ArtifactSkillPoints();
-            if (infinityData.goalSetter > 0) totalEarned += infinityData.goalSetter;
-            int spent = CountAssignedSkillPoints();
-            long recalculated = totalEarned - spent;
-            skillTreeData.skillPointsTree = recalculated < 0 ? 0 : recalculated;
-            saveSettings.hasFixedIP = true;
-        }
-
-        private int CountAssignedSkillPoints()
-        {
-            GameDataRegistry registry = GameDataRegistry.Instance;
-            if (registry == null || registry.skillDatabase == null) return 0;
-
-            int total = 0;
-            foreach (SkillDefinition skill in registry.skillDatabase.skills)
-            {
-                if (skill == null || string.IsNullOrEmpty(skill.id)) continue;
-                if (IsSkillOwned(skill.id)) total += skill.cost;
-            }
-
-            return total;
-        }
-
-        private void LoadDictionaries()
-        {
-            foreach (KeyValuePair<int, SkillTreeItem> variable in SkillTree)
-            {
-                infinityData.SkillTreeSaveData ??= new Dictionary<int, bool>();
-                infinityData.SkillTreeSaveData.TryAdd(variable.Key, variable.Value.Owned);
-                variable.Value.Owned = infinityData.SkillTreeSaveData[variable.Key];
-            }
-        }
-
-        [TabGroup("SaveData", "Buttons"), Button]
-        public void SaveToClipboard()
-        {
-            SaveDictionaries();
-            byte[] fullBytes = SirenixSerializationUtility.SerializeValue(saveSettings, DataFormat.JSON);
-            SaveDataSettings snapshot = SaveSnapshotBuilder.CreateSaveSnapshotForStorage(
-                saveSettings,
-                includeBase64Fields: false,
-                buildOwnedBitsetFromRuntime: BuildOwnedBitsetFromRuntime,
-                getAutoAssignmentSkillIds: GetAutoAssignmentSkillIds);
-            if (beta)
-            {
-                byte[] bytes = SirenixSerializationUtility.SerializeValue(snapshot, DataFormat.JSON);
-                string compactJson = Encoding.UTF8.GetString(bytes);
-                GUIUtility.systemCopyBuffer = compactJson;
-                string fullClipboardJson = Encoding.UTF8.GetString(fullBytes);
-                Debug.Log(
-                    $"SaveToClipboard size: {fullBytes.Length} -> {bytes.Length} bytes (raw JSON), " +
-                    $"{fullClipboardJson.Length} -> {compactJson.Length} chars (clipboard).");
-                return;
-            }
-
-            byte[] binaryBytes = SirenixSerializationUtility.SerializeValue(snapshot, DataFormat.Binary);
-            string compactClipboard = SaveCodec.EncodeBinary(binaryBytes, compress: true);
-            GUIUtility.systemCopyBuffer = compactClipboard;
-            string fullClipboardBinary = SaveCodec.EncodeBinary(SirenixSerializationUtility.SerializeValue(saveSettings, DataFormat.Binary),
-                compress: true);
-            Debug.Log(
-                $"SaveToClipboard size: {fullBytes.Length} -> {binaryBytes.Length} bytes (binary), " +
-                $"{fullClipboardBinary.Length} -> {compactClipboard.Length} chars (clipboard).");
-        }
-
-        [TabGroup("SaveData", "Buttons"), Button]
-        public void ExportSaveDebugJson()
-        {
-            SaveDictionaries();
-            byte[] fullBytes = SirenixSerializationUtility.SerializeValue(saveSettings, DataFormat.JSON);
-            SaveDataSettings snapshot = SaveSnapshotBuilder.CreateSaveSnapshotForStorage(
-                saveSettings,
-                includeBase64Fields: false,
-                buildOwnedBitsetFromRuntime: BuildOwnedBitsetFromRuntime,
-                getAutoAssignmentSkillIds: GetAutoAssignmentSkillIds);
-            byte[] bytes = SirenixSerializationUtility.SerializeValue(snapshot, DataFormat.Binary);
-            byte[] compressed = SaveCodec.CompressBytes(bytes);
-            var dto = new ExportSaveDto
-            {
-                version = CurrentSaveVersion,
-                format = "binary+gzip+base64",
-                rawBytes = bytes.Length,
-                compressedBytes = compressed.Length,
-                data = Convert.ToBase64String(compressed)
-            };
-            string json = JsonUtility.ToJson(dto);
-
-            string folderPath = GetSaveDebugFolderPath();
-            Directory.CreateDirectory(folderPath);
-            string filePath = Path.Combine(folderPath, $"save-debug-{DateTime.UtcNow:yyyyMMdd-HHmmss}.json");
-            File.WriteAllText(filePath, json, Encoding.UTF8);
-
-            Debug.Log($"Save debug JSON written to: {filePath}");
-            string debugFullClipboard = beta ? Encoding.UTF8.GetString(fullBytes) : SaveCodec.EncodeBinary(fullBytes, compress: true);
-            string debugCompactClipboard = beta ? json : SaveCodec.EncodeBinary(bytes, compress: true);
-            Debug.Log(
-                $"ExportSaveDebugJson size: {fullBytes.Length} -> {bytes.Length} bytes (raw JSON), " +
-                $"{debugFullClipboard.Length} -> {debugCompactClipboard.Length} chars (clipboard).");
-#if UNITY_EDITOR
-            AssetDatabase.Refresh();
-#endif
-        }
-
-        [TabGroup("SaveData", "Buttons"), Button]
-        public void PreviewSkillPointRecalc()
-        {
-            SkillPointRecalcResult result = CalculateSkillPointRecalc();
-            LogSkillPointRecalc(result, apply: false);
-        }
-
-        [TabGroup("SaveData", "Buttons"), Button]
-        public void ApplySkillPointRecalc()
-        {
-            SkillPointRecalcResult result = CalculateSkillPointRecalc();
-            skillTreeData.skillPointsTree = result.Recalculated;
-            LogSkillPointRecalc(result, apply: true);
-            UpdateSkills?.Invoke();
-        }
-
-        private SkillPointRecalcResult CalculateSkillPointRecalc()
-        {
-            long earnedKnown = 0;
-            if (prestigeData != null) earnedKnown += prestigeData.permanentSkillPoint;
-            earnedKnown += ArtifactSkillPoints();
-            if (infinityData != null && infinityData.goalSetter > 0) earnedKnown += infinityData.goalSetter;
-
-            int spent = CountAssignedSkillPoints();
-            long recalculated = earnedKnown - spent;
-            if (recalculated < 0) recalculated = 0;
-
-            return new SkillPointRecalcResult(earnedKnown, spent, skillTreeData.skillPointsTree, recalculated);
-        }
-
-        private void LogSkillPointRecalc(SkillPointRecalcResult result, bool apply)
-        {
-            string note =
-                "Note: This uses known sources (permanent IP, artifact points, goalSetter) and ignores points awarded" +
-                " by research/debug/manual adjustments.";
-            string action = apply ? "Applied" : "Preview";
-            Debug.Log(
-                $"{action} SkillPointRecalc -> earnedKnown: {result.EarnedKnown}, spent: {result.Spent}, " +
-                $"current: {result.Current}, recalculated: {result.Recalculated}. {note}");
-        }
-
-        private readonly struct SkillPointRecalcResult
-        {
-            public long EarnedKnown { get; }
-            public int Spent { get; }
-            public long Current { get; }
-            public long Recalculated { get; }
-
-            public SkillPointRecalcResult(long earnedKnown, int spent, long current, long recalculated)
-            {
-                EarnedKnown = earnedKnown;
-                Spent = spent;
-                Current = current;
-                Recalculated = recalculated;
-            }
-        }
-
-        private string GetSaveDebugFolderPath()
-        {
-#if UNITY_EDITOR
-            string projectRoot = Path.GetFullPath(Path.Combine(Application.dataPath, ".."));
-            return Path.Combine(projectRoot, "Documentation", "savedebugging");
-#else
-            return Path.Combine(Application.persistentDataPath, "savedebugging");
-#endif
-        }
-
-        private void SaveDictionaries()
-        {
-            infinityData.SkillTreeSaveData = new Dictionary<int, bool>();
-            if (infinityData.skillStateById != null && infinityData.skillStateById.Count > 0)
-            {
-                foreach (KeyValuePair<string, SkillState> entry in infinityData.skillStateById)
-                {
-                    if (SkillIdMap.TryGetLegacyKey(entry.Key, out int key))
-                    {
-                        infinityData.SkillTreeSaveData[key] = entry.Value.owned;
-                    }
-                }
-
-                return;
-            }
-
-            if (infinityData.skillOwnedById != null && infinityData.skillOwnedById.Count > 0)
-            {
-                foreach (KeyValuePair<string, bool> entry in infinityData.skillOwnedById)
-                {
-                    if (SkillIdMap.TryGetLegacyKey(entry.Key, out int key))
-                    {
-                        infinityData.SkillTreeSaveData[key] = entry.Value;
-                    }
-                }
-
-                return;
-            }
-
-            foreach (KeyValuePair<int, SkillTreeItem> value in SkillTree)
-            {
-                if (!infinityData.SkillTreeSaveData.ContainsKey(value.Key))
-                    infinityData.SkillTreeSaveData.Add(value.Key, value.Value.Owned);
-                infinityData.SkillTreeSaveData[value.Key] = value.Value.Owned;
-            }
-        }
-
-        #endregion
-
-        private void SetSaveReady(bool ready)
-        {
-            _isSaveReady = ready;
-            if (_isSaveReady) ScheduleAutoSave();
-        }
-
-        public IDisposable SuppressPresetSync()
-        {
-            _suppressPresetSyncDepth++;
-            return new PresetSyncScope(this);
-        }
-
-        private sealed class PresetSyncScope : IDisposable
-        {
-            private Oracle _owner;
-
-            public PresetSyncScope(Oracle owner)
-            {
-                _owner = owner;
-            }
-
-            public void Dispose()
-            {
-                if (_owner == null) return;
-                _owner._suppressPresetSyncDepth = Mathf.Max(0, _owner._suppressPresetSyncDepth - 1);
-                _owner = null;
-            }
-        }
-
-        private bool IsPresetSyncSuppressed()
-        {
-            return _suppressPresetSyncDepth > 0;
-        }
-
-        private void ScheduleAutoSave()
-        {
-            if (_autoSaveScheduled) return;
-            InvokeRepeating(nameof(Save), 60, 60);
-            _autoSaveScheduled = true;
-        }
-
-        private void ScheduleQuickSave()
-        {
-            if (!_isSaveReady) return;
-            // Debounce rapid UI interactions (skill assign/unassign, sliders, etc.)
-            CancelInvoke(nameof(QuickSave));
-            Invoke(nameof(QuickSave), QuickSaveDelaySeconds);
-            _quickSaveScheduled = true;
-        }
-
-        private void QuickSave()
-        {
-            _quickSaveScheduled = false;
-            SaveInternal(false);
-        }
-
-        public void SyncSelectedPresetAutoAssignFromCurrent()
-        {
-            if (IsPresetSyncSuppressed()) return;
-
-            DysonVerseSaveData data = saveSettings?.dysonVerseSaveData;
-            if (data == null) return;
-
-            int presetIndex = Mathf.Clamp(data.selectedPreset, 1, 5);
-            // Clone to avoid sharing list references between "current" and slot storage.
-            List<string> ids = new List<string>(GetAutoAssignmentSkillIds());
-            SetPresetAutoAssignmentSkillIds(presetIndex, ids);
-            SyncSelectedPresetBotDistributionFromCurrent();
-        }
-
-        public void SyncSelectedPresetBotDistributionFromCurrent()
-        {
-            if (IsPresetSyncSuppressed()) return;
-
-            DysonVerseSaveData data = saveSettings?.dysonVerseSaveData;
-            if (data == null) return;
-
-            int presetIndex = Mathf.Clamp(data.selectedPreset, 1, 5);
-            switch (presetIndex)
-            {
-                case 1:
-                    data.botDistPreset1 = prestigeData.botDistribution;
-                    break;
-                case 2:
-                    data.botDistPreset2 = prestigeData.botDistribution;
-                    break;
-                case 3:
-                    data.botDistPreset3 = prestigeData.botDistribution;
-                    break;
-                case 4:
-                    data.botDistPreset4 = prestigeData.botDistribution;
-                    break;
-                case 5:
-                    data.botDistPreset5 = prestigeData.botDistribution;
-                    break;
-            }
-
-            ScheduleQuickSave();
-        }
-
-	        private void ApplyMigrations()
+	        private void EnsureSkillOwnershipData()
 	        {
-	            if (saveSettings == null) return;
-
-	            MigrationRegistry registry = BuildMigrationRegistry();
-            if (registry.LatestVersion != CurrentSaveVersion)
-            {
-                string message =
-                    $"Migration registry latest version {registry.LatestVersion} does not match CurrentSaveVersion {CurrentSaveVersion}.";
-#if UNITY_EDITOR
-                Debug.LogError(message);
-                throw new InvalidOperationException(message);
-#else
-                Debug.LogWarning(message);
-#endif
-	            }
-
-	            MigrationRunOptions options = BuildMigrationOptions(false);
-	            // Transactional migration: run against a deep copy and only commit on success.
-	            SaveDataSettings original = saveSettings;
-	            SaveDataSettings working = (SaveDataSettings)SirenixSerializationUtility.CreateCopy(original);
-	            saveSettings = working;
-	            try
-	            {
-	                MigrationRunResult result = MigrationRunner.Run(this, registry, options);
-	                Debug.Log(result.ToReportString());
-	            }
-	            catch
-	            {
-	                saveSettings = original;
-	                throw;
-	            }
-	        }
-
-        public MigrationRunResult RunMigrationDryRun()
-        {
-            if (saveSettings == null) return null;
-
-            MigrationRegistry registry = BuildMigrationRegistry();
-            MigrationRunOptions options = BuildMigrationOptions(true);
-            options.ThrowOnError = false;
-            MigrationRunResult result = MigrationRunner.Run(this, registry, options);
-            Debug.Log(result.ToReportString());
-            return result;
-        }
-
-        public MigrationRunResult RunMigrationDryRun(SaveDataSettings saveData)
-        {
-            if (saveData == null) return null;
-
-            SaveDataSettings original = saveSettings;
-            try
-            {
-                saveSettings = saveData;
-                return RunMigrationDryRun();
-            }
-            finally
-            {
-                saveSettings = original;
-            }
-        }
-
-        private MigrationRunOptions BuildMigrationOptions(bool dryRun)
-        {
-            return new MigrationRunOptions
-            {
-                DryRun = dryRun,
-                CaptureSnapshots = dryRun,
-                IncludeEnsureStep = true,
-                ThrowOnError = !dryRun,
-                UpdateLastSuccessfulLoadUtc = !dryRun,
-                EnsureAction = _ =>
-                {
-                    EnsureSkillOwnershipData();
-                    EnsureSkillAutoAssignmentIds();
-                    EnsureResearchLevelData();
-                    EnsurePackedSettingsFlags();
-                    EnsureInfinitySparseArrays();
-                }
-            };
-        }
-
-	        private MigrationRegistry BuildMigrationRegistry()
-	        {
-	            var registry = new MigrationRegistry();
-	            registry.AddStep(new MigrationStep(
-	                targetVersion: 11,
-	                name: "Consolidated migration (V11)",
-	                summary: "Upgrade any legacy save to V11 in a single step (skipping intermediate chains).",
-	                apply: _ => { MigrateToV11(); }));
-
-	            return registry;
-	        }
-
-	        private void MigrateToV11()
-	        {
-	            // Apply all prior migrations idempotently in one pass.
-	            MigrateSkillOwnershipToIds();
-	            MigrateSkillAutoAssignmentIds();
-	            MigrateResearchLevelsToIds();
-	            MigrateSkillStateToIds();
-	            MigrateAvocadoData();
-	            MigrateMegaStructureData();
-	            MigrateSkillBitsets();
-	            PackSettingsFlags();
-	            MigratePresetAutoAssignOrder();
-	            MigrateAvotationProgress();
-	        }
-
-        private void MigrateAvotationProgress()
-        {
-            if (saveSettings == null) return;
-
-            saveSettings.avotationProgressStep = Mathf.Clamp(saveSettings.avotationProgressStep, 0, 7);
-
-            // Legacy complete flag remains canonical for the +4 skill-point reward.
-            if (saveSettings.avotation)
-            {
-                saveSettings.avotationProgressStep = 7;
-            }
-            else if (saveSettings.avotationProgressStep >= 7)
-            {
-                saveSettings.avotation = true;
-                saveSettings.avotationProgressStep = 7;
-            }
-        }
-
-        private void EnsureSkillOwnershipData()
-        {
-            if (infinityData == null) return;
+	            if (infinityData == null) return;
             EnsureSkillStateData();
             EnsureSkillOwnedBitset();
             SyncSkillOwnedByIdFromBitset();
@@ -3438,33 +2918,10 @@ namespace Expansion
             SaveInternal(false);
         }
 
-	        public void AttemptSaveRecovery()
-	        {
-	            string canonicalPath = SavePaths.GetCanonicalSavePath();
-	            if (File.Exists(canonicalPath))
-	            {
-	                string text = File.ReadAllText(canonicalPath, Encoding.UTF8);
-	                GUIUtility.systemCopyBuffer = text;
-	                recoveryText.text = "Save Copied";
-	                Debug.Log("Copied canonical save to clipboard");
-	                return;
-	            }
-
-	            string legacyOdinPath = Path.Combine(Application.persistentDataPath, fileName + ".idsOdin");
-	            if (File.Exists(legacyOdinPath))
-	            {
-	                byte[] bytes = File.ReadAllBytes(legacyOdinPath);
-	                Debug.Log("Copied to clipboard");
-
-	                GUIUtility.systemCopyBuffer = beta ? Encoding.UTF8.GetString(bytes) : Convert.ToBase64String(bytes);
-	                recoveryText.text = "Save Copied";
-	            }
-	        }
-
-		        public void SaveState()
-		        {
-		            SaveDictionaries();
-		            PackSettingsFlags();
+			        public void SaveState()
+			        {
+			            SaveDictionaries();
+			            PackSettingsFlags();
 		            SaveDataSettings snapshot = SaveSnapshotBuilder.CreateSaveSnapshotForStorage(
 		                saveSettings,
 		                includeBase64Fields: false,
